final test cleanup
