+++ apollo-federation/src/operation/mod.rs
@@ -921,17 +921,6 @@ impl FragmentSpreadSelection {
         })
     }
 
-    pub(crate) fn from_fragment(
-        fragment: &Node<Fragment>,
-        directives: &executable::DirectiveList,
-    ) -> Self {
-        let spread = FragmentSpread::from_fragment(fragment, directives);
-        Self {
-            spread,
-            selection_set: fragment.selection_set.clone(),
-        }
-    }
-
     /// Creates a fragment spread selection (in an optimized operation).
     /// - `named_fragments`: Named fragment definitions that are rebased for the element's schema.
     pub(crate) fn new(
@@ -2171,15 +2160,6 @@ impl SelectionSet {
         })
     }
 
-    /// In a normalized selection set containing only fields and inline fragments,
-    /// iterate over all the fields that may be selected.
-    ///
-    /// # Preconditions
-    /// The selection set must not contain named fragment spreads.
-    pub(crate) fn field_selections(&self) -> FieldSelectionsIter<'_> {
-        FieldSelectionsIter::new(self.selections.values())
-    }
-
     /// # Preconditions
     /// The selection set must not contain named fragment spreads.
     fn fields_in_set(&self) -> Vec<CollectedFieldInSet> {
@@ -2317,36 +2297,6 @@ impl<'a> IntoIterator for &'a SelectionSet {
 
     fn into_iter(self) -> Self::IntoIter {
         self.selections.values()
-    }
-}
-
-pub(crate) struct FieldSelectionsIter<'sel> {
-    stack: Vec<selection_map::Values<'sel>>,
-}
-
-impl<'sel> FieldSelectionsIter<'sel> {
-    fn new(iter: selection_map::Values<'sel>) -> Self {
-        Self { stack: vec![iter] }
-    }
-}
-
-impl<'sel> Iterator for FieldSelectionsIter<'sel> {
-    type Item = &'sel Arc<FieldSelection>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.stack.last_mut()?.next() {
-            None if self.stack.len() == 1 => None,
-            None => {
-                self.stack.pop();
-                self.next()
-            }
-            Some(Selection::Field(field)) => Some(field),
-            Some(Selection::InlineFragment(frag)) => {
-                self.stack.push(frag.selection_set.selections.values());
-                self.next()
-            }
-            Some(Selection::FragmentSpread(_frag)) => unreachable!(),
-        }
     }
 }
 
@@ -2625,16 +2575,6 @@ impl Field {
     pub(crate) fn parent_type_position(&self) -> CompositeTypeDefinitionPosition {
         self.field_position.parent()
     }
-
-    pub(crate) fn types_can_be_merged(&self, other: &Self) -> Result<bool, FederationError> {
-        let self_definition = self.field_position.get(self.schema().schema())?;
-        let other_definition = other.field_position.get(self.schema().schema())?;
-        types_can_be_merged(
-            &self_definition.ty,
-            &other_definition.ty,
-            self.schema().schema(),
-        )
-    }
 }
 
 impl InlineFragmentSelection {
@@ -2732,23 +2672,6 @@ impl InlineFragmentSelection {
         ))
     }
 
-    /// Construct a new InlineFragmentSelection out of a selection set.
-    /// - The new type condition will be the same as the selection set's type.
-    pub(crate) fn from_selection_set(
-        parent_type_position: CompositeTypeDefinitionPosition,
-        selection_set: SelectionSet,
-        directives: DirectiveList,
-    ) -> Self {
-        let inline_fragment_data = InlineFragment {
-            schema: selection_set.schema.clone(),
-            parent_type_position,
-            type_condition_position: selection_set.type_position.clone().into(),
-            directives,
-            selection_id: SelectionId::new(),
-        };
-        InlineFragmentSelection::new(inline_fragment_data, selection_set)
-    }
-
     pub(crate) fn casted_type(&self) -> &CompositeTypeDefinitionPosition {
         self.inline_fragment
             .type_condition_position
@@ -2811,29 +2734,8 @@ impl NamedFragments {
         NamedFragments::initialize_in_dependency_order(fragments, schema)
     }
 
-    pub(crate) fn is_empty(&self) -> bool {
-        self.fragments.len() == 0
-    }
-
-    pub(crate) fn len(&self) -> usize {
-        self.fragments.len()
-    }
-
     pub(crate) fn iter(&self) -> impl Iterator<Item = &Node<Fragment>> {
         self.fragments.values()
-    }
-
-    pub(crate) fn iter_rev(&self) -> impl Iterator<Item = &Node<Fragment>> {
-        self.fragments.values().rev()
-    }
-
-    pub(crate) fn iter_mut(&mut self) -> indexmap::map::IterMut<'_, Name, Node<Fragment>> {
-        Arc::make_mut(&mut self.fragments).iter_mut()
-    }
-
-    // Calls `retain` on the underlying `IndexMap`.
-    pub(crate) fn retain(&mut self, mut predicate: impl FnMut(&Name, &Node<Fragment>) -> bool) {
-        Arc::make_mut(&mut self.fragments).retain(|name, fragment| predicate(name, fragment));
     }
 
     fn insert(&mut self, fragment: Fragment) {
@@ -2928,32 +2830,6 @@ impl NamedFragments {
                 *current_count += 1;
             }
         })
-    }
-
-    /// When we rebase named fragments on a subgraph schema, only a subset of what the fragment handles may belong
-    /// to that particular subgraph. And there are a few sub-cases where that subset is such that we basically need or
-    /// want to consider to ignore the fragment for that subgraph, and that is when:
-    /// 1. the subset that apply is actually empty. The fragment wouldn't be valid in this case anyway.
-    /// 2. the subset is a single leaf field: in that case, using the one field directly is just shorter than using
-    ///    the fragment, so we consider the fragment don't really apply to that subgraph. Technically, using the
-    ///    fragment could still be of value if the fragment name is a lot smaller than the one field name, but it's
-    ///    enough of a niche case that we ignore it. Note in particular that one sub-case of this rule that is likely
-    ///    to be common is when the subset ends up being just `__typename`: this would basically mean the fragment
-    ///    don't really apply to the subgraph, and that this will ensure this is the case.
-    pub(crate) fn is_selection_set_worth_using(selection_set: &SelectionSet) -> bool {
-        if selection_set.selections.len() == 0 {
-            return false;
-        }
-        if selection_set.selections.len() == 1 {
-            // true if NOT field selection OR non-leaf field
-            return if let Some(Selection::Field(field_selection)) = selection_set.selections.first()
-            {
-                field_selection.selection_set.is_some()
-            } else {
-                true
-            };
-        }
-        true
     }
 }
