+++ apollo-federation/src/operation/optimize.rs
@@ -81,26 +81,6 @@ impl Selection {
         }
         Ok(None)
     }
-
-    /// Computes the set-intersection of self and other
-    /// - If there are respective sub-selections, then we compute their intersections and add them
-    ///   (if not empty).
-    /// - Otherwise, the intersection is same as `self`.
-    fn intersection(&self, other: &Selection) -> Result<Option<Selection>, FederationError> {
-        if let (Some(self_sub_selection), Some(other_sub_selection)) =
-            (self.selection_set(), other.selection_set())
-        {
-            let common = self_sub_selection.intersection(other_sub_selection)?;
-            if common.is_empty() {
-                return Ok(None);
-            } else {
-                return self
-                    .with_updated_selections(self_sub_selection.type_position.clone(), common)
-                    .map(Some);
-            }
-        }
-        Ok(Some(self.clone()))
-    }
 }
 
 impl SelectionSet {
@@ -115,35 +95,6 @@ impl SelectionSet {
                     v.minus(other_v)
                 } else {
                     Ok(Some(v.clone()))
-                }
-            })
-            .collect::<Result<Vec<_>, _>>()? // early break in case of Err
-            .into_iter()
-            .flatten();
-        Ok(SelectionSet::from_raw_selections(
-            self.schema.clone(),
-            self.type_position.clone(),
-            iter,
-        ))
-    }
-
-    /// Computes the set-intersection of self and other
-    fn intersection(&self, other: &SelectionSet) -> Result<SelectionSet, FederationError> {
-        if self.is_empty() {
-            return Ok(self.clone());
-        }
-        if other.is_empty() {
-            return Ok(other.clone());
-        }
-
-        let iter = self
-            .selections
-            .values()
-            .map(|v| {
-                if let Some(other_v) = other.selections.get(v.key()) {
-                    v.intersection(other_v)
-                } else {
-                    Ok(None)
                 }
             })
             .collect::<Result<Vec<_>, _>>()? // early break in case of Err
@@ -219,25 +170,6 @@ impl Operation {
         generator.visit_selection_set(&mut self.selection_set)?;
         self.named_fragments = generator.into_inner();
         Ok(())
-    }
-
-    // PORT_NOTE: This mirrors the JS version's `Operation.expandAllFragments`. But this method is
-    // mainly for unit tests. The actual port of `expandAllFragments` is in `normalize_operation`.
-    #[cfg(test)]
-    fn expand_all_fragments_and_normalize(&self) -> Result<Self, FederationError> {
-        let selection_set = self
-            .selection_set
-            .expand_all_fragments()?
-            .flatten_unnecessary_fragments(
-                &self.selection_set.type_position,
-                &self.named_fragments,
-                &self.schema,
-            )?;
-        Ok(Self {
-            named_fragments: Default::default(),
-            selection_set,
-            ..self.clone()
-        })
     }
 }
 
@@ -421,27 +353,8 @@ impl FragmentGenerator {
 
 #[cfg(test)]
 mod tests {
-    use apollo_compiler::ExecutableDocument;
-
     use super::*;
     use crate::operation::tests::*;
-
-    macro_rules! assert_without_fragments {
-        ($operation: expr, @$expected: literal) => {{
-            let without_fragments = $operation.expand_all_fragments_and_normalize().unwrap();
-            insta::assert_snapshot!(without_fragments, @$expected);
-            without_fragments
-        }};
-    }
-
-    macro_rules! assert_optimized {
-        ($operation: expr, $named_fragments: expr, @$expected: literal) => {{
-            let mut optimized = $operation.clone();
-            optimized.reuse_fragments(&$named_fragments).unwrap();
-            validate_operation(&$operation.schema, &optimized.to_string());
-            insta::assert_snapshot!(optimized, @$expected)
-        }};
-    }
 
     /// Returns a consistent GraphQL name for the given index.
     fn fragment_name(mut index: usize) -> Name {
@@ -474,1619 +387,6 @@ mod tests {
         assert_eq!(fragment_name(0), "a");
         assert_eq!(fragment_name(100), "Vb");
         assert_eq!(fragment_name(usize::MAX), "oS5Uz8g3Iqw");
-    }
-
-    #[test]
-    fn duplicate_fragment_spreads_after_fragment_expansion() {
-        // This is a regression test for FED-290, making sure `make_select` method can handle
-        // duplicate fragment spreads.
-        // During optimization, `make_selection` may merge multiple fragment spreads with the same
-        // key. This can happen in the case below where `F1` and `F2` are expanded and generating
-        // two duplicate `F_shared` spreads in the definition of `fragment F_target`.
-        let schema_doc = r#"
-            type Query {
-                t: T
-                t2: T
-            }
-
-            type T {
-                id: ID!
-                a: Int!
-                b: Int!
-                c: Int!
-            }
-        "#;
-
-        let query = r#"
-            fragment F_shared on T {
-                id
-                a
-            }
-            fragment F1 on T {
-                ...F_shared
-                b
-            }
-
-            fragment F2 on T {
-                ...F_shared
-                c
-            }
-
-            fragment F_target on T {
-                ...F1
-                ...F2
-            }
-
-            query {
-                t {
-                    ...F_target
-                }
-                t2 {
-                    ...F_target
-                }
-            }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema_doc), query);
-        let expanded = operation.expand_all_fragments_and_normalize().unwrap();
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-        fragment F_target on T {
-          id
-          a
-          b
-          c
-        }
-
-        {
-          t {
-            ...F_target
-          }
-          t2 {
-            ...F_target
-          }
-        }
-        "###);
-    }
-
-    #[test]
-    fn optimize_fragments_using_other_fragments_when_possible() {
-        let schema = r#"
-              type Query {
-                t: I
-              }
-
-              interface I {
-                b: Int
-                u: U
-              }
-
-              type T1 implements I {
-                a: Int
-                b: Int
-                u: U
-              }
-
-              type T2 implements I {
-                x: String
-                y: String
-                b: Int
-                u: U
-              }
-
-              union U = T1 | T2
-        "#;
-
-        let query = r#"
-              fragment OnT1 on T1 {
-                a
-                b
-              }
-
-              fragment OnT2 on T2 {
-                x
-                y
-              }
-
-              fragment OnI on I {
-                b
-              }
-
-              fragment OnU on U {
-                ...OnI
-                ...OnT1
-                ...OnT2
-              }
-
-              query {
-                t {
-                  ...OnT1
-                  ...OnT2
-                  ...OnI
-                  u {
-                    ...OnU
-                  }
-                }
-              }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema), query);
-
-        let expanded = assert_without_fragments!(
-            operation,
-            @r###"
-        {
-          t {
-            ... on T1 {
-              a
-              b
-            }
-            ... on T2 {
-              x
-              y
-            }
-            b
-            u {
-              ... on I {
-                b
-              }
-              ... on T1 {
-                a
-                b
-              }
-              ... on T2 {
-                x
-                y
-              }
-            }
-          }
-        }
-        "###
-        );
-
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-              fragment OnU on U {
-                ... on I {
-                  b
-                }
-                ... on T1 {
-                  a
-                  b
-                }
-                ... on T2 {
-                  x
-                  y
-                }
-              }
-
-              {
-                t {
-                  ...OnU
-                  u {
-                    ...OnU
-                  }
-                }
-              }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragments_using_other_fragments() {
-        let schema = r#"
-              type Query {
-                t: I
-              }
-
-              interface I {
-                b: Int
-                c: Int
-                u1: U
-                u2: U
-              }
-
-              type T1 implements I {
-                a: Int
-                b: Int
-                c: Int
-                me: T1
-                u1: U
-                u2: U
-              }
-
-              type T2 implements I {
-                x: String
-                y: String
-                b: Int
-                c: Int
-                u1: U
-                u2: U
-              }
-
-              union U = T1 | T2
-        "#;
-
-        let query = r#"
-              fragment OnT1 on T1 {
-                a
-                b
-              }
-
-              fragment OnT2 on T2 {
-                x
-                y
-              }
-
-              fragment OnI on I {
-                b
-                c
-              }
-
-              fragment OnU on U {
-                ...OnI
-                ...OnT1
-                ...OnT2
-              }
-
-              query {
-                t {
-                  ...OnT1
-                  ...OnT2
-                  u1 {
-                    ...OnU
-                  }
-                  u2 {
-                    ...OnU
-                  }
-                  ... on T1 {
-                    me {
-                      ...OnI
-                    }
-                  }
-                }
-              }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema), query);
-
-        let expanded = assert_without_fragments!(
-            &operation,
-            @r###"
-              {
-                t {
-                  ... on T1 {
-                    a
-                    b
-                    me {
-                      b
-                      c
-                    }
-                  }
-                  ... on T2 {
-                    x
-                    y
-                  }
-                  u1 {
-                    ... on I {
-                      b
-                      c
-                    }
-                    ... on T1 {
-                      a
-                      b
-                    }
-                    ... on T2 {
-                      x
-                      y
-                    }
-                  }
-                  u2 {
-                    ... on I {
-                      b
-                      c
-                    }
-                    ... on T1 {
-                      a
-                      b
-                    }
-                    ... on T2 {
-                      x
-                      y
-                    }
-                  }
-                }
-              }
-        "###);
-
-        // We should reuse and keep all fragments, because 1) onU is used twice and 2)
-        // all the other ones are used once in the query, and once in onU definition.
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-              fragment OnT1 on T1 {
-                a
-                b
-              }
-
-              fragment OnT2 on T2 {
-                x
-                y
-              }
-
-              fragment OnI on I {
-                b
-                c
-              }
-
-              fragment OnU on U {
-                ...OnI
-                ...OnT1
-                ...OnT2
-              }
-
-              {
-                t {
-                  ... on T1 {
-                    ...OnT1
-                    me {
-                      ...OnI
-                    }
-                  }
-                  ...OnT2
-                  u1 {
-                    ...OnU
-                  }
-                  u2 {
-                    ...OnU
-                  }
-                }
-              }
-        "###);
-    }
-
-    macro_rules! test_fragments_roundtrip {
-        ($schema_doc: expr, $query: expr, @$expanded: literal) => {{
-            let schema = parse_schema($schema_doc);
-            let operation = parse_operation(&schema, $query);
-            let without_fragments = operation.expand_all_fragments_and_normalize().unwrap();
-            insta::assert_snapshot!(without_fragments, @$expanded);
-
-            let mut optimized = without_fragments;
-            optimized.reuse_fragments(&operation.named_fragments).unwrap();
-            validate_operation(&operation.schema, &optimized.to_string());
-            assert_eq!(optimized.to_string(), operation.to_string());
-        }};
-    }
-
-    /// Tests ported from JS codebase rely on special behavior of
-    /// `Operation::reuse_fragments_for_roundtrip_test` that is specific for testing, since it makes it
-    /// easier to write tests.
-    macro_rules! test_fragments_roundtrip_legacy {
-        ($schema_doc: expr, $query: expr, @$expanded: literal) => {{
-            let schema = parse_schema($schema_doc);
-            let operation = parse_operation(&schema, $query);
-            let without_fragments = operation.expand_all_fragments_and_normalize().unwrap();
-            insta::assert_snapshot!(without_fragments, @$expanded);
-
-            let mut optimized = without_fragments;
-            optimized.reuse_fragments_for_roundtrip_test(&operation.named_fragments).unwrap();
-            validate_operation(&operation.schema, &optimized.to_string());
-            assert_eq!(optimized.to_string(), operation.to_string());
-        }};
-    }
-
-    #[test]
-    fn handles_fragments_with_nested_selections() {
-        let schema_doc = r#"
-              type Query {
-                t1a: T1
-                t2a: T1
-              }
-
-              type T1 {
-                t2: T2
-              }
-
-              type T2 {
-                x: String
-                y: String
-              }
-        "#;
-
-        let query = r#"
-                fragment OnT1 on T1 {
-                  t2 {
-                    x
-                  }
-                }
-
-                query {
-                  t1a {
-                    ...OnT1
-                    t2 {
-                      y
-                    }
-                  }
-                  t2a {
-                    ...OnT1
-                  }
-                }
-        "#;
-
-        test_fragments_roundtrip!(schema_doc, query, @r###"
-                {
-                  t1a {
-                    t2 {
-                      x
-                      y
-                    }
-                  }
-                  t2a {
-                    t2 {
-                      x
-                    }
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn handles_nested_fragments_with_field_intersection() {
-        let schema_doc = r#"
-            type Query {
-                t: T
-            }
-
-            type T {
-                a: A
-                b: Int
-            }
-
-            type A {
-                x: String
-                y: String
-                z: String
-            }
-        "#;
-
-        // The subtlety here is that `FA` contains `__typename` and so after we're reused it, the
-        // selection will look like:
-        // {
-        //   t {
-        //     a {
-        //       ...FA
-        //     }
-        //   }
-        // }
-        // But to recognize that `FT` can be reused from there, we need to be able to see that
-        // the `__typename` that `FT` wants is inside `FA` (and since FA applies on the parent type `A`
-        // directly, it is fine to reuse).
-        let query = r#"
-            fragment FA on A {
-                __typename
-                x
-                y
-            }
-
-            fragment FT on T {
-                a {
-                __typename
-                ...FA
-                }
-            }
-
-            query {
-                t {
-                ...FT
-                }
-            }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-        {
-          t {
-            a {
-              __typename
-              x
-              y
-            }
-          }
-        }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragment_matching_subset_of_field_selection() {
-        let schema_doc = r#"
-              type Query {
-                t: T
-              }
-
-              type T {
-                a: String
-                b: B
-                c: Int
-                d: D
-              }
-
-              type B {
-                x: String
-                y: String
-              }
-
-              type D {
-                m: String
-                n: String
-              }
-        "#;
-
-        let query = r#"
-                fragment FragT on T {
-                  b {
-                    __typename
-                    x
-                  }
-                  c
-                  d {
-                    m
-                  }
-                }
-
-                {
-                  t {
-                    ...FragT
-                    d {
-                      n
-                    }
-                    a
-                  }
-                }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  t {
-                    b {
-                      __typename
-                      x
-                    }
-                    c
-                    d {
-                      m
-                      n
-                    }
-                    a
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragment_matching_subset_of_inline_fragment_selection() {
-        // Pretty much the same test than the previous one, but matching inside a fragment selection inside
-        // of inside a field selection.
-        // PORT_NOTE: ` implements I` was added in the definition of `type T`, so that validation can pass.
-        let schema_doc = r#"
-          type Query {
-            i: I
-          }
-
-          interface I {
-            a: String
-          }
-
-          type T implements I {
-            a: String
-            b: B
-            c: Int
-            d: D
-          }
-
-          type B {
-            x: String
-            y: String
-          }
-
-          type D {
-            m: String
-            n: String
-          }
-        "#;
-
-        let query = r#"
-            fragment FragT on T {
-              b {
-                __typename
-                x
-              }
-              c
-              d {
-                m
-              }
-            }
-
-            {
-              i {
-                ... on T {
-                  ...FragT
-                  d {
-                    n
-                  }
-                  a
-                }
-              }
-            }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-            {
-              i {
-                ... on T {
-                  b {
-                    __typename
-                    x
-                  }
-                  c
-                  d {
-                    m
-                    n
-                  }
-                  a
-                }
-              }
-            }
-        "###);
-    }
-
-    #[test]
-    fn intersecting_fragments() {
-        let schema_doc = r#"
-              type Query {
-                t: T
-              }
-
-              type T {
-                a: String
-                b: B
-                c: Int
-                d: D
-              }
-
-              type B {
-                x: String
-                y: String
-              }
-
-              type D {
-                m: String
-                n: String
-              }
-        "#;
-
-        // Note: the code that reuse fragments iterates on fragments in the order they are defined
-        // in the document, but when it reuse a fragment, it puts it at the beginning of the
-        // selection (somewhat random, it just feel often easier to read), so the net effect on
-        // this example is that `Frag2`, which will be reused after `Frag1` will appear first in
-        // the re-optimized selection. So we put it first in the input too so that input and output
-        // actually match (the `testFragmentsRoundtrip` compares strings, so it is sensible to
-        // ordering; we could theoretically use `Operation.equals` instead of string equality,
-        // which wouldn't really on ordering, but `Operation.equals` is not entirely trivial and
-        // comparing strings make problem a bit more obvious).
-        let query = r#"
-                fragment Frag1 on T {
-                  b {
-                    x
-                  }
-                  c
-                  d {
-                    m
-                  }
-                }
-
-                fragment Frag2 on T {
-                  a
-                  b {
-                    __typename
-                    x
-                  }
-                  d {
-                    m
-                    n
-                  }
-                }
-
-                {
-                  t {
-                    ...Frag1
-                    ...Frag2
-                  }
-                }
-        "#;
-
-        // PORT_NOTE: `__typename` and `x`'s placements are switched in Rust.
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  t {
-                    b {
-                      __typename
-                      x
-                    }
-                    c
-                    d {
-                      m
-                      n
-                    }
-                    a
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn fragments_application_makes_type_condition_trivial() {
-        let schema_doc = r#"
-              type Query {
-                t: T
-              }
-
-              interface I {
-                x: String
-              }
-
-              type T implements I {
-                x: String
-                a: String
-              }
-        "#;
-
-        let query = r#"
-                fragment FragI on I {
-                  x
-                  ... on T {
-                    a
-                  }
-                }
-
-                {
-                  t {
-                    ...FragI
-                  }
-                }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  t {
-                    x
-                    a
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragment_matching_at_the_top_level_of_another_fragment() {
-        let schema_doc = r#"
-              type Query {
-                t: T
-              }
-
-              type T {
-                a: String
-                u: U
-              }
-
-              type U {
-                x: String
-                y: String
-              }
-        "#;
-
-        let query = r#"
-                fragment Frag1 on T {
-                  a
-                }
-
-                fragment Frag2 on T {
-                  u {
-                    x
-                    y
-                  }
-                  ...Frag1
-                }
-
-                fragment Frag3 on Query {
-                  t {
-                    ...Frag2
-                  }
-                }
-
-                {
-                  ...Frag3
-                }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  t {
-                    u {
-                      x
-                      y
-                    }
-                    a
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragments_used_in_context_where_they_get_trimmed() {
-        let schema_doc = r#"
-              type Query {
-                t1: T1
-              }
-
-              interface I {
-                x: Int
-              }
-
-              type T1 implements I {
-                x: Int
-                y: Int
-              }
-
-              type T2 implements I {
-                x: Int
-                z: Int
-              }
-        "#;
-
-        let query = r#"
-                fragment FragOnI on I {
-                  ... on T1 {
-                    y
-                  }
-                  ... on T2 {
-                    z
-                  }
-                }
-
-                {
-                  t1 {
-                    ...FragOnI
-                  }
-                }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  t1 {
-                    y
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragments_used_in_the_context_of_non_intersecting_abstract_types() {
-        let schema_doc = r#"
-              type Query {
-                i2: I2
-              }
-
-              interface I1 {
-                x: Int
-              }
-
-              interface I2 {
-                y: Int
-              }
-
-              interface I3 {
-                z: Int
-              }
-
-              type T1 implements I1 & I2 {
-                x: Int
-                y: Int
-              }
-
-              type T2 implements I1 & I3 {
-                x: Int
-                z: Int
-              }
-        "#;
-
-        let query = r#"
-                fragment FragOnI1 on I1 {
-                  ... on I2 {
-                    y
-                  }
-                  ... on I3 {
-                    z
-                  }
-                }
-
-                {
-                  i2 {
-                    ...FragOnI1
-                  }
-                }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  i2 {
-                    ... on I1 {
-                      ... on I2 {
-                        y
-                      }
-                      ... on I3 {
-                        z
-                      }
-                    }
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn handles_fragments_on_union_in_context_with_limited_intersection() {
-        let schema_doc = r#"
-              type Query {
-                t1: T1
-              }
-
-              union U = T1 | T2
-
-              type T1 {
-                x: Int
-              }
-
-              type T2 {
-                y: Int
-              }
-        "#;
-
-        let query = r#"
-                fragment OnU on U {
-                  ... on T1 {
-                    x
-                  }
-                  ... on T2 {
-                    y
-                  }
-                }
-
-                {
-                  t1 {
-                    ...OnU
-                  }
-                }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                {
-                  t1 {
-                    x
-                  }
-                }
-        "###);
-    }
-
-    #[test]
-    fn off_by_1_error() {
-        let schema = r#"
-              type Query {
-                t: T
-              }
-              type T {
-                id: String!
-                a: A
-                v: V
-              }
-              type A {
-                id: String!
-              }
-              type V {
-                t: T!
-              }
-        "#;
-
-        let query = r#"
-              {
-                t {
-                  ...TFrag
-                  v {
-                    t {
-                      id
-                      a {
-                        __typename
-                        id
-                      }
-                    }
-                  }
-                }
-              }
-
-              fragment TFrag on T {
-                __typename
-                id
-              }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema), query);
-
-        let expanded = assert_without_fragments!(
-            operation,
-            @r###"
-              {
-                t {
-                  __typename
-                  id
-                  v {
-                    t {
-                      id
-                      a {
-                        __typename
-                        id
-                      }
-                    }
-                  }
-                }
-              }
-            "###
-        );
-
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-        fragment TFrag on T {
-          __typename
-          id
-        }
-
-        {
-          t {
-            ...TFrag
-            v {
-              t {
-                ...TFrag
-                a {
-                  __typename
-                  id
-                }
-              }
-            }
-          }
-        }
-        "###);
-    }
-
-    #[test]
-    fn removes_all_unused_fragments() {
-        let schema = r#"
-              type Query {
-                t1: T1
-              }
-
-              union U1 = T1 | T2 | T3
-              union U2 =      T2 | T3
-
-              type T1 {
-                x: Int
-              }
-
-              type T2 {
-                y: Int
-              }
-
-              type T3 {
-                z: Int
-              }
-        "#;
-
-        let query = r#"
-              query {
-                t1 {
-                  ...Outer
-                }
-              }
-
-              fragment Outer on U1 {
-                ... on T1 {
-                  x
-                }
-                ... on T2 {
-                  ... Inner
-                }
-                ... on T3 {
-                  ... Inner
-                }
-              }
-
-              fragment Inner on U2 {
-                ... on T2 {
-                  y
-                }
-              }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema), query);
-
-        let expanded = assert_without_fragments!(
-            operation,
-            @r###"
-              {
-                t1 {
-                  x
-                }
-              }
-            "###
-        );
-
-        // This is a bit of contrived example, but the reusing code will be able
-        // to figure out that the `Outer` fragment can be reused and will initially
-        // do so, but it's only use once, so it will expand it, which yields:
-        // {
-        //   t1 {
-        //     ... on T1 {
-        //       x
-        //     }
-        //     ... on T2 {
-        //       ... Inner
-        //     }
-        //     ... on T3 {
-        //       ... Inner
-        //     }
-        //   }
-        // }
-        // and so `Inner` will not be expanded (it's used twice). Except that
-        // the `flatten_unnecessary_fragments` code is apply then and will _remove_ both instances
-        // of `.... Inner`. Which is ok, but we must make sure the fragment
-        // itself is removed since it is not used now, which this test ensures.
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-              {
-                t1 {
-                  x
-                }
-              }
-        "###);
-    }
-
-    #[test]
-    fn removes_fragments_only_used_by_unused_fragments() {
-        // Similar to the previous test, but we artificially add a
-        // fragment that is only used by the fragment that is finally
-        // unused.
-        let schema = r#"
-              type Query {
-                t1: T1
-              }
-
-              union U1 = T1 | T2 | T3
-              union U2 =      T2 | T3
-
-              type T1 {
-                x: Int
-              }
-
-              type T2 {
-                y1: Y
-                y2: Y
-              }
-
-              type T3 {
-                z: Int
-              }
-
-              type Y {
-                v: Int
-              }
-        "#;
-
-        let query = r#"
-              query {
-                t1 {
-                  ...Outer
-                }
-              }
-
-              fragment Outer on U1 {
-                ... on T1 {
-                  x
-                }
-                ... on T2 {
-                  ... Inner
-                }
-                ... on T3 {
-                  ... Inner
-                }
-              }
-
-              fragment Inner on U2 {
-                ... on T2 {
-                  y1 {
-                    ...WillBeUnused
-                  }
-                  y2 {
-                    ...WillBeUnused
-                  }
-                }
-              }
-
-              fragment WillBeUnused on Y {
-                v
-              }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema), query);
-
-        let expanded = assert_without_fragments!(
-            operation,
-            @r###"
-              {
-                t1 {
-                  x
-                }
-              }
-            "###
-        );
-
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-              {
-                t1 {
-                  x
-                }
-              }
-        "###);
-    }
-
-    #[test]
-    fn keeps_fragments_used_by_other_fragments() {
-        let schema = r#"
-              type Query {
-                t1: T
-                t2: T
-              }
-
-              type T {
-                a1: Int
-                a2: Int
-                b1: B
-                b2: B
-              }
-
-              type B {
-                x: Int
-                y: Int
-              }
-        "#;
-
-        let query = r#"
-              query {
-                t1 {
-                  ...TFields
-                }
-                t2 {
-                  ...TFields
-                }
-              }
-
-              fragment TFields on T {
-                ...DirectFieldsOfT
-                b1 {
-                  ...BFields
-                }
-                b2 {
-                  ...BFields
-                }
-              }
-
-              fragment DirectFieldsOfT on T {
-                a1
-                a2
-              }
-
-              fragment BFields on B {
-                x
-                y
-              }
-        "#;
-
-        let operation = parse_operation(&parse_schema(schema), query);
-
-        let expanded = assert_without_fragments!(
-            operation,
-            @r###"
-              {
-                t1 {
-                  a1
-                  a2
-                  b1 {
-                    x
-                    y
-                  }
-                  b2 {
-                    x
-                    y
-                  }
-                }
-                t2 {
-                  a1
-                  a2
-                  b1 {
-                    x
-                    y
-                  }
-                  b2 {
-                    x
-                    y
-                  }
-                }
-              }
-            "###
-        );
-
-        // The `DirectFieldsOfT` fragments should not be kept as it is used only once within `TFields`,
-        // but the `BFields` one should be kept.
-        assert_optimized!(expanded, operation.named_fragments, @r###"
-        fragment BFields on B {
-          x
-          y
-        }
-
-        fragment TFields on T {
-          a1
-          a2
-          b1 {
-            ...BFields
-          }
-          b2 {
-            ...BFields
-          }
-        }
-
-        {
-          t1 {
-            ...TFields
-          }
-          t2 {
-            ...TFields
-          }
-        }
-        "###);
-    }
-
-    ///
-    /// applied directives
-    ///
-
-    #[test]
-    fn reuse_fragments_with_same_directive_in_the_fragment_selection() {
-        let schema_doc = r#"
-                type Query {
-                  t1: T
-                  t2: T
-                  t3: T
-                }
-
-                type T {
-                  a: Int
-                  b: Int
-                  c: Int
-                  d: Int
-                }
-        "#;
-
-        let query = r#"
-                  fragment DirectiveInDef on T {
-                    a @include(if: $cond1)
-                  }
-
-                  query myQuery($cond1: Boolean!, $cond2: Boolean!) {
-                    t1 {
-                      a
-                    }
-                    t2 {
-                      ...DirectiveInDef
-                    }
-                    t3 {
-                      a @include(if: $cond2)
-                    }
-                  }
-        "#;
-
-        test_fragments_roundtrip_legacy!(schema_doc, query, @r###"
-                  query myQuery($cond1: Boolean!, $cond2: Boolean!) {
-                    t1 {
-                      a
-                    }
-                    t2 {
-                      a @include(if: $cond1)
-                    }
-                    t3 {
-                      a @include(if: $cond2)
-                    }
-                  }
-        "###);
-    }
-
-    #[test]
-    fn reuse_fragments_with_directives_on_inline_fragments() {
-        let schema_doc = r#"
-                type Query {
-                  t1: T
-                  t2: T
-                  t3: T
-                }
-
-                type T {
-                  a: Int
-                  b: Int
-                  c: Int
-                  d: Int
-                }
-        "#;
-
-        let query = r#"
-                  fragment NoDirectiveDef on T {
-                    a
-                  }
-
-                  query myQuery($cond1: Boolean!) {
-                    t1 {
-                      ...NoDirectiveDef
-                    }
-                    t2 {
-                      ...NoDirectiveDef @include(if: $cond1)
-                    }
-                  }
-        "#;
-
-        test_fragments_roundtrip!(schema_doc, query, @r###"
-                  query myQuery($cond1: Boolean!) {
-                    t1 {
-                      a
-                    }
-                    t2 {
-                      ... on T @include(if: $cond1) {
-                        a
-                      }
-                    }
-                  }
-        "###);
-    }
-
-    #[test]
-    fn reuse_fragments_with_directive_on_typename() {
-        let schema = r#"
-            type Query {
-              t1: T
-              t2: T
-              t3: T
-            }
-
-            type T {
-              a: Int
-              b: Int
-              c: Int
-              d: Int
-            }
-        "#;
-        let query = r#"
-            query A ($if: Boolean!) {
-              t1 { b a ...x }
-              t2 { ...x }
-            }
-            query B {
-              # Because this inline fragment is exactly the same shape as `x`,
-              # except for a `__typename` field, it may be tempting to reuse it.
-              # But `x.__typename` has a directive with a variable, and this query
-              # does not have that variable declared, so it can't be used.
-              t3 { ... on T { a c } }
-            }
-            fragment x on T {
-                __typename @include(if: $if)
-                a
-                c
-            }
-        "#;
-        let schema = parse_schema(schema);
-        let query = ExecutableDocument::parse_and_validate(schema.schema(), query, "query.graphql")
-            .unwrap();
-
-        let operation_a =
-            Operation::from_operation_document(schema.clone(), &query, Some("A")).unwrap();
-        let operation_b =
-            Operation::from_operation_document(schema.clone(), &query, Some("B")).unwrap();
-        let expanded_b = operation_b.expand_all_fragments_and_normalize().unwrap();
-
-        assert_optimized!(expanded_b, operation_a.named_fragments, @r###"
-        query B {
-          t3 {
-            a
-            c
-          }
-        }
-        "###);
-    }
-
-    #[test]
-    fn reuse_fragments_with_non_intersecting_types() {
-        let schema = r#"
-            type Query {
-              t: T
-              s: S
-              s2: S
-              i: I
-            }
-
-            interface I {
-                a: Int
-                b: Int
-            }
-
-            type T implements I {
-              a: Int
-              b: Int
-
-              c: Int
-              d: Int
-            }
-            type S implements I {
-              a: Int
-              b: Int
-
-              f: Int
-              g: Int
-            }
-        "#;
-        let query = r#"
-            query A ($if: Boolean!) {
-              t { ...x }
-              s { ...x }
-              i { ...x }
-            }
-            query B {
-              s {
-                # this matches fragment x once it is flattened,
-                # because the `...on T` condition does not intersect with our
-                # current type `S`
-                __typename
-                a b
-              }
-              s2 {
-                # same snippet to get it to use the fragment
-                __typename
-                a b
-              }
-            }
-            fragment x on I {
-                __typename
-                a
-                b
-                ... on T { c d @include(if: $if) }
-            }
-        "#;
-        let schema = parse_schema(schema);
-        let query = ExecutableDocument::parse_and_validate(schema.schema(), query, "query.graphql")
-            .unwrap();
-
-        let operation_a =
-            Operation::from_operation_document(schema.clone(), &query, Some("A")).unwrap();
-        let operation_b =
-            Operation::from_operation_document(schema.clone(), &query, Some("B")).unwrap();
-        let expanded_b = operation_b.expand_all_fragments_and_normalize().unwrap();
-
-        assert_optimized!(expanded_b, operation_a.named_fragments, @r###"
-        query B {
-          s {
-            __typename
-            a
-            b
-          }
-          s2 {
-            __typename
-            a
-            b
-          }
-        }
-        "###);
     }
 
     ///
